requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
crashtest==0.3.1
cryptography==37.0.4
distlib==0.3.6
dulwich==0.20.50
filelock==3.9.0
frozenlist==1.3.3
html5lib==1.1
idna==3.4
importlib-metadata==5.2.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
magic-filter==1.0.9
more-itertools==9.0.0
msgpack==1.0.4
multidict==6.0.4
packaging==22.0
pexpect==4.8.0
pipenv==2022.12.19
pkginfo==1.9.2
platformdirs==2.6.2
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.4
pylev==1.4.0
pyrsistent==0.19.3
python-dotenv==0.21.0
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
tomlkit==0.11.6
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
yarl==1.8.2
zipp==3.11.0

loguru~=0.6.0
peewee~=3.15.4
aiogram~=3.0.0b6